Move world into evm_arithmetization
