+++ evm_arithmetization/Cargo.toml
@@ -15,8 +15,13 @@ homepage.workspace = true
 keywords.workspace = true
 
 [dependencies]
+alloy.workspace = true
+alloy-compat.workspace = true
 anyhow.workspace = true
+bitvec.workspace = true
 bytes.workspace = true
+copyvec.workspace = true
+either.workspace = true
 env_logger.workspace = true
 ethereum-types.workspace = true
 hashbrown.workspace = true
@@ -43,7 +48,7 @@ serde = { workspace = true, features = ["derive"] }
 serde-big-array.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
-smt_trie = { workspace = true, optional = true }
+smt_trie = { workspace = true }
 starky = { workspace = true, features = ["parallel"] }
 static_assertions.workspace = true
 thiserror.workspace = true
@@ -51,6 +56,7 @@ tiny-keccak.workspace = true
 tokio.workspace = true
 tower-lsp = "0.20.0"
 tracing.workspace = true
+u4.workspace = true
 url.workspace = true
 zk_evm_common.workspace = true
 zk_evm_proc_macro.workspace = true
@@ -64,7 +70,7 @@ ripemd.workspace = true
 default = ["eth_mainnet"]
 asmtools = ["hex"]
 polygon_pos = []
-cdk_erigon = ["smt_trie"]
+cdk_erigon = []
 eth_mainnet = []
 
 [[bin]]
