+++ trace_decoder/src/lib.rs
@@ -62,14 +62,10 @@ const _DEVELOPER_DOCS: () = ();
 mod interface;
 
 pub use interface::*;
-use keccak_hash::H256;
-use smt_trie::code::hash_bytecode_h256;
 
-mod tries;
 mod type1;
 mod type2;
 mod wire;
-mod world;
 
 pub use core::{entrypoint, WireDisposition};
 
@@ -100,27 +96,6 @@ mod hex {
             None => T::from_hex(&*s),
         }
         .map_err(D::Error::custom)
-    }
-}
-
-/// Utility trait to leverage a specific hash function across Type1 and Type2
-/// zkEVM variants.
-pub(crate) trait Hasher {
-    fn hash(bytes: &[u8]) -> H256;
-}
-
-pub(crate) struct PoseidonHash;
-pub(crate) struct KeccakHash;
-
-impl Hasher for PoseidonHash {
-    fn hash(bytes: &[u8]) -> H256 {
-        hash_bytecode_h256(bytes)
-    }
-}
-
-impl Hasher for KeccakHash {
-    fn hash(bytes: &[u8]) -> H256 {
-        keccak_hash::keccak(bytes)
     }
 }
